receiver/requirements.txt
paho-mqtt
influxdb-client
influxdb==5.3.1
